move try_download into output.py
