+++ tools/wheel_resolver/BUILD
@@ -14,6 +14,7 @@ python_library(
     srcs = [
         "__init__.py",
         "wheel.py",
+        "output.py",
     ],
     deps = [
         "//third_party/python:click",
