+++ tools/wheel_resolver/__init__.py
@@ -1,36 +1,17 @@
 import click
 import typing
 import requests
-import os
-import urllib.request
 import logging
 import click_log
 import sys
 import tools.wheel_resolver.wheel as wheel
+import tools.wheel_resolver.output as output
 import packaging.tags as tags
 import distlib.locators
 
 _LOGGER = logging.getLogger(__name__)
 
 click_log.basic_config(_LOGGER)
-
-def try_download(url):
-    """
-    Try to download url to $OUTS. Returns false if
-    it failed.
-    """
-    output = os.environ.get("OUTS")
-    if output is None:
-        _LOGGER.critical("No output directory found")
-        sys.exit(1)
-
-    try:
-        urllib.request.urlretrieve(url, output)
-    except urllib.error.HTTPError:
-        return False
-
-    return True
-
 
 @click.command()
 @click.option(
@@ -126,7 +107,7 @@ def main(
         )
         sys.exit(1)
 
-    if not try_download(u):
+    if not output.try_download(u):
         _LOGGER.error("Could not download from %r", u)
         sys.exit(1)
